Added two methods to insert business accounts
These two methods work with the info entered on the form to add business classes. I would be nice to query for the existing account types to select from, however it does not do this for now.
Also, since the database accepts a 'Y' or 'N' for the approved column, I added a get for the character. I know that returning booleans is more normal, however I think that would add an extra step here.