fix issue with a null pointer exception.
I originally set up the wine table to allow nulls in the tasting notes and promo materials. However the Java Bean doesn't like that. I did put a null check on the .jsp page so it will not throw the error page. I also filled the item that was coming back null with an empty string.